eval: move template format to a new module
