+++ lib/template/Makefile.am
@@ -7,6 +7,7 @@ templateinclude_HEADERS = \
 	lib/template/templates.h		\
 	lib/template/macros.h			\
 	lib/template/function.h			\
+	lib/template/eval.h			\
 	lib/template/simple-function.h		\
 	lib/template/repr.h			\
 	lib/template/compiler.h			\
@@ -17,6 +18,7 @@ templateinclude_HEADERS = \
 template_sources = \
 	lib/template/templates.c		\
 	lib/template/macros.c			\
+	lib/template/eval.c			\
 	lib/template/simple-function.c		\
 	lib/template/repr.c			\
 	lib/template/compiler.c			\
