+++ lib/template/templates.h
@@ -26,14 +26,11 @@
 #define TEMPLATES_H_INCLUDED
 
 #include "syslog-ng.h"
-#include "common-template-typedefs.h"
+#include "eval.h"
 #include "timeutils/zoneinfo.h"
 #include "type-hinting.h"
+#include "common-template-typedefs.h"
 #include "atomic.h"
-
-#define LTZ_LOCAL 0
-#define LTZ_SEND  1
-#define LTZ_MAX   2
 
 #define LOG_TEMPLATE_ERROR log_template_error_quark()
 
@@ -55,7 +52,7 @@ typedef enum
 } LogTemplateOnError;
 
 /* structure that represents an expandable syslog-ng template */
-typedef struct _LogTemplate
+struct _LogTemplate
 {
   GAtomicCounter ref_cnt;
   gchar *name;
@@ -64,27 +61,6 @@ typedef struct _LogTemplate
   GlobalConfig *cfg;
   guint escape:1, def_inline:1, trivial:1;
   TypeHint type_hint;
-} LogTemplate;
-
-/* template expansion options that can be influenced by the user and
- * is static throughout the runtime for a given configuration. There
- * are call-site specific options too, those are specified as
- * arguments to log_template_format() */
-struct _LogTemplateOptions
-{
-  gboolean initialized;
-  /* timestamp format as specified by ts_format() */
-  gint ts_format;
-  /* number of digits in the fraction of a second part, specified using frac_digits() */
-  gint frac_digits;
-  gboolean use_fqdn;
-
-  /* timezone for LTZ_LOCAL/LTZ_SEND settings */
-  gchar *time_zone[LTZ_MAX];
-  TimeZoneInfo *time_zone_info[LTZ_MAX];
-
-  /* Template error handling settings */
-  gint on_error;
 };
 
 /* appends the formatted output into result */
@@ -95,20 +71,11 @@ gboolean log_template_compile(LogTemplate *self, const gchar *template, GError *
 void log_template_compile_literal_string(LogTemplate *self, const gchar *literal);
 gboolean log_template_is_trivial(LogTemplate *self);
 const gchar *log_template_get_trivial_value(LogTemplate *self, LogMessage *msg, gssize *value_len);
-void log_template_format(LogTemplate *self, LogMessage *lm, const LogTemplateOptions *opts, gint tz, gint32 seq_num,
-                         const gchar *context_id, GString *result);
-void log_template_append_format(LogTemplate *self, LogMessage *lm, const LogTemplateOptions *opts, gint tz,
-                                gint32 seq_num, const gchar *context_id, GString *result);
-void log_template_append_format_with_context(LogTemplate *self, LogMessage **messages, gint num_messages,
-                                             const LogTemplateOptions *opts, gint tz, gint32 seq_num, const gchar *context_id, GString *result);
-void log_template_format_with_context(LogTemplate *self, LogMessage **messages, gint num_messages,
-                                      const LogTemplateOptions *opts, gint tz, gint32 seq_num, const gchar *context_id, GString *result);
 void log_template_set_name(LogTemplate *self, const gchar *name);
 
 LogTemplate *log_template_new(GlobalConfig *cfg, const gchar *name);
 LogTemplate *log_template_ref(LogTemplate *s);
 void log_template_unref(LogTemplate *s);
-
 
 void log_template_options_init(LogTemplateOptions *options, GlobalConfig *cfg);
 void log_template_options_destroy(LogTemplateOptions *options);
